organize files in /utils
