+++ src/utils/index.js
@@ -1,26 +1,22 @@
 import $ from 'jquery'
-import { storage } from './storage'
 import { links   } from './links'
 
 export { env              } from './env'
-export { search           } from './search'
-export { user, userAction } from './user'
-export { handleData       } from './data'
 export { api              } from './http'
-export { prompt           } from './js/CTPrompt'
+export { user             } from './user'
+export { search           } from './search'
+export { userAction       } from './useraction'
 export { CTEpubGenerator  } from './epub-gen'
+export { CTPrompt, prompt } from './prompt'
+export { CTPreference     } from './user-preference'
 
 export * from './constants'
-
-export { CTPreference } from './js/CTPreference'
-export { CTPrompt } from './js/CTPrompt'
 
 
 /**
  * Objects for switching pages and storing some general functions
  */
 export const util = {
-  ...storage,
   links: links,
 
   refresh: function() {
