+++ src/utils/prompt/CTPrompt.js
@@ -1,72 +1,17 @@
 import _ from 'lodash'
-import { v4 as uuid4 } from 'uuid'
-import { util } from 'utils'
+
+// import stylesheet of prompt element
 import 'components/stylesheets/ct-prompt.css'
 
+import {
+  PROMPT_ID,
+  createPromptBoxElem,
+  createPromptElem,
+} from './prompt-creators'
 
-const PROMPT_ID = 'ct___prompt'
-
-function createPromptBoxElem(text, options) {
-  let { status, contact, refresh, onClose } = options
-  // Prompt Box
-  const promptBoxEl = document.createElement('div')
-  promptBoxEl.className = 'ct-prompt-box ' + status
-  promptBoxEl.id = 'ctp-box-' + uuid4()
-
-  // Text
-  const txtEl = document.createElement('div')
-  txtEl.className = 'ct-prompt-text'
-  txtEl.innerText = text
-  promptBoxEl.appendChild(txtEl)
-
-  // Contact Link
-  if (contact) {
-    const contactLinkEl = document.createElement('a')
-    contactLinkEl.className = 'ct-prompt-link'
-    contactLinkEl.href = util.links.contactUs()
-    contactLinkEl.innerText = 'CONTACT US'
-    promptBoxEl.appendChild(contactLinkEl)
-  }
-  // Refresh Link
-  if (refresh) {
-    const refreshLinkEl = document.createElement('a')
-    refreshLinkEl.className = 'ct-prompt-link'
-    refreshLinkEl.href = window.location.pathname
-    refreshLinkEl.innerHTML = `<i class="material-icons">refresh</i><span>REFRESH</span>`
-    
-    promptBoxEl.appendChild(refreshLinkEl)
-  }
-
-  // Close Button
-  const closeBtnEl = document.createElement('button')
-  closeBtnEl.className = 'plain-btn ct-prompt-close-btn'
-  closeBtnEl.innerHTML = '<i class="material-icons">close</i>'
-  if (typeof onClose === "function") {
-    closeBtnEl.addEventListener('click', e => {
-      onClose(promptBoxEl.id)
-    })
-  }
-  promptBoxEl.appendChild(closeBtnEl)
-
-  return promptBoxEl
-}
-
-function createPromptElem(text, options) {
-  let { position } = options
-  let promptBoxEl = createPromptBoxElem(text, options)
-  // Prompt
-  const promptEl = document.createElement('div')
-  promptEl.id = PROMPT_ID
-  promptEl.appendChild(promptBoxEl)
-  // promptEl.setAttribute('data-position', position)
-  promptEl.className = position
-  // insert prompt elem after root
-  const rootEl = document.getElementById('root')
-  rootEl.parentNode.insertBefore(promptEl, rootEl.nextSibling)
-
-  return promptBoxEl
-}
-
+/**
+ * Class used to create and handle prompt component
+ */
 export class CTPrompt {
   constructor() {
     this.promptIds = []
@@ -82,6 +27,7 @@ export class CTPrompt {
   /**
    * Close a prompt box
    * @param {String} boxId - the id of the prompt box; close all prompts if undefined
+   * @returns {void}
    */
   close(boxId) {
     if (boxId === undefined) {
@@ -102,6 +48,7 @@ export class CTPrompt {
 
   /**
    * Close all the prompt boxes
+   * @returns {void}
    */
   closeAll() {
     const closePrompt = this.close
@@ -124,6 +71,8 @@ export class CTPrompt {
    * - in 'primary', 'success', 'error' (default 'primary')
    * @property position - The position of the prompt
    * - in 'bottom left', 'bottom right' (default 'bottom right')
+   * 
+   * @returns {void}
    */
   addOne(prompt={
     text: '',
@@ -187,10 +136,10 @@ export class CTPrompt {
    * - in 'primary', 'success', 'error' (default 'primary')
    * @property position - The position of the prompt
    * - in 'bottom left', 'bottom right' (default 'bottom right')
+   * 
+   * @returns {void}
    */
   addMany(prompts=[]) {
     _.forEach(prompts, prp => this.addOne(prp))
   }
 }
-
-export const prompt = new CTPrompt()
