+++ src/utils/prompt/index.js
@@ -1,0 +1,5 @@
+import { CTPrompt } from './CTPrompt'
+
+export { CTPrompt } from './CTPrompt'
+
+export const prompt = new CTPrompt()
