--- src/utils/storage/watch-options.js
@@ -1,7 +1,0 @@
-import { isBrowser } from 'react-device-detect'
-
-export const watchOptionsStorage = {
-  playbackrateKey: 'watch-options-plbk-rate',
-
-  
-}
