+++ src/utils/user-preference/CTPreference.js
@@ -14,6 +14,7 @@ export class CTPreference {
    * @param {String} key The key to the localStorage value
    * @param {Boolean} bool The value to set
    * @param {Boolean} setfalse if true and bool is false, set the value to be 'false' instead of removing from localStorage
+   * @returns {string} the value stored in localStorage if no bool provided
    */
   localStorage(key, bool, setfalse=false) {
     if (bool === undefined) return this[key]
@@ -32,6 +33,7 @@ export class CTPreference {
   /**
    * Function used to determine if an elem in localStorage is true
    * @param {String} key The key to the localStorage value
+   * @returns {Boolean} if an elem in localStorage is true
    */
   isTrue(key) {
     return localStorage.getItem(key) === this.TRUE
@@ -40,6 +42,7 @@ export class CTPreference {
   /**
    * Function used to determine if an elem in localStorage is false
    * @param {String} key The key to the localStorage value
+   * @returns {Boolean} if an elem in localStorage is false
    */
   isFalse(key) {
     return localStorage.getItem(key) === this.FALSE
